+++ src/neo/SmartContract/ApplicationEngine.Runtime.cs
@@ -27,75 +27,6 @@ namespace Neo.SmartContract
         public static readonly InteropDescriptor System_Runtime_Notify = Register("System.Runtime.Notify", nameof(RuntimeNotify), 0_01000000, TriggerType.All, CallFlags.AllowNotify);
         public static readonly InteropDescriptor System_Runtime_GetNotifications = Register("System.Runtime.GetNotifications", nameof(GetNotifications), 0_00010000, TriggerType.All, CallFlags.None);
         public static readonly InteropDescriptor System_Runtime_GasLeft = Register("System.Runtime.GasLeft", nameof(GasLeft), 0_00000400, TriggerType.All, CallFlags.None);
-        public static readonly InteropDescriptor CreateCallback = Register("System.Runtime.CreateCallback", nameof(Runtime_CreateCallback), 0_00000400, TriggerType.All, CallFlags.None);
-        public static readonly InteropDescriptor CreateSyscallCallback = Register("System.Runtime.CreateSyscallCallback", nameof(Runtime_CreateSyscallCallback), 0_00000400, TriggerType.All, CallFlags.None);
-        public static readonly InteropDescriptor InvokeCallback = Register("System.Runtime.InvokeCallback", nameof(Runtime_InvokeCallback), 0_00000400, TriggerType.All, CallFlags.None);
-
-        internal void Runtime_InvokeCallback(Callback callback)
-        {
-            callback.Action(this);
-        }
-
-        internal void Runtime_CreateCallback(Pointer pointer, int parcount)
-        {
-            // Save arguments
-
-            var arguments = new StackItem[parcount];
-            for (int x = parcount - 1; x >= 0; x--)
-            {
-                arguments[x] = Pop();
-            }
-
-            // Push callback
-
-            Push(new InteropInterface(new Callback(engine =>
-            {
-                // Clone context
-
-                var newContext = CurrentContext.Clone(0);
-                newContext.InstructionPointer = pointer.Position;
-
-                // Copy arguments
-
-                foreach (var arg in arguments)
-                {
-                    newContext.EvaluationStack.Push(arg);
-                }
-
-                // Load context
-
-                LoadContext(newContext);
-            },
-            arguments)));
-        }
-
-        internal void Runtime_CreateSyscallCallback(uint method, int parcount)
-        {
-            // Save arguments
-
-            var arguments = new StackItem[parcount];
-            for (int x = parcount - 1; x >= 0; x--)
-            {
-                arguments[x] = Pop();
-            }
-
-            // Push callback
-
-            Push(new InteropInterface(new Callback(engine =>
-            {
-                // Copy arguments
-
-                foreach (var arg in arguments)
-                {
-                    Push(arg);
-                }
-
-                // Execute syscall
-
-                OnSysCall(method);
-            },
-            arguments)));
-        }
 
         private static bool CheckItemForNotification(StackItem state)
         {
